Add scripts for generating BkBookImages.prp
ExtractTexturesPRP.py
Use this script to extract any PRP's textures to a folder losslessly.
This means any DXT mipmaps will be DDS files, and any JPEGs or PNGs will
be saved out as-is.

MakeImageLibJSON.py
This is a bit of a hack, but run it on a PRP file containing a
plImageLibMod, and it will generate a JSON file matching the output of
ExtractTexturesPRP.py... mostly.

MakeImageLibPRP.py
This one gets the job done. Convert a JSON file and a bunch of images to
a PRP file with a plImageLibMod.